chore: rebalance test batches
